Fix nsqlookupd template to handle empty attribute
The nsqlookupd broadcast_address attribute defaults to an empty string
in this cookbook. This results in an empty --broadcast-address option
in the upstart config, thus setting its value to the next line (--tombstone-lifetime)
and generally screwing things up.

This template fix follows the convention from the nsqd default upstart template.
It only sets the option if the attribute is not empty.